requirements.txt
PyQt5==5.15.11
python-vlc==3.0.18121
mutagen==1.46.0
requests==2.28.1